0.3.0: New Methods & Breaking Changes
The `next_with_buffer()` method now returns a boolean instead of
a result. Instead of returning a result on an error, it will now
panic instead, as any error here is caused by programmer error,
and should be treated as such. The assertion check is only enabled
during debug builds, so release builds will not pay for the checks
anymore.

Additionally, the `nth()` method is now specialized; documentation
has been improved; and the `get_index()` and `set_index()` methods
have been implemented. The new methods will allow for obtain the
state of the iterator, and setting the state, at any given point.